requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
boto3==1.36.17
openai>=1.0.0
streamlit==1.38.0
requests==2.32.3